Create middleware who checks can edit or can delete

+++ routes/api.php
@@ -20,8 +20,8 @@ Route::post('/login', [LoginController::class, 'login']);
 Route::middleware('auth:sanctum')->prefix('tasks')->group(function () {
     Route::get('', [TaskController::class, 'index'])->name('task.index');
     Route::post('', [TaskController::class, 'store'])->name('task.store');
-    Route::put('/{id}', [TaskController::class, 'update'])->name('task.update');
-    Route::delete('/{id}', [TaskController::class, 'destroy'])->name('task.destroy');
+    Route::put('/{id}', [TaskController::class, 'update'])->name('task.update')->middleware('permission.edit');
+    Route::delete('/{id}', [TaskController::class, 'destroy'])->name('task.destroy')->middleware('permission.delete');
     Route::put('/set-performed/{id}', [TaskController::class, 'set_performed'])->name('task.set_performed');
 
 });
